spore/chaterelle/pyproject.toml
[project]
name = "chaterelle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.1.1",
    "huggingface-hub[hf-xet]>=0.32.4",
    "openai-whisper>=20240930",
    "python-dotenv>=1.1.0",
    "soundfile>=0.13.1",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]
